Fix issues in synapse expressions
Fix issues in synapse expressions
Fixes wso2/product-micro-integrator/issues/3872
Fixes wso2/product-micro-integrator/issues/3874

+++ modules/core/src/main/java/org/apache/synapse/util/synapse/expression/ast/LiteralNode.java
@@ -72,7 +72,7 @@ public class LiteralNode implements ExpressionNode {
             return new ExpressionResult(Integer.parseInt(value));
         } catch (NumberFormatException e1) {
             try {
-                return new ExpressionResult(Float.parseFloat(value));
+                return new ExpressionResult(Long.parseLong(value));
             } catch (NumberFormatException e2) {
                 try {
                     return new ExpressionResult(Double.parseDouble(value));
